move format types to module
